plugins/ircd: add
This plugin adds primitive support for acting like an UnrealIRCd server,
making it possible to send messages on the bridge and receive messages
on IRC.

The fact that this acts like a server allows it to do all sorts of neat
stuff, like acting like multiple users (each client on each instance
gets its own user).

Things left to implement include handling service messages, but this
rudimentary support is useful on its own.