core/bridge: add support for more than two nodes on a route
Consider a hypothetical bridge with 10 interconnected nodes.
Previously, we required the configuration to contain 10C2 routes, one
for every pair of connected nodes. This is not so bad if you have four
nodes, but at ten instances this leads to 45 route entries.

This commit adds support for specifying more than two nodes in a route,
which are all considered to be interconnected. I.e.,

```python
routes = [["a", "b", "c"], ["c", "d"]]
```

will route messages between the triangle `a`, `b`, `c` (new),
and between `c` and `d` (unchanged).